pyproject.toml
[project]
name = "roasted"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai==1.7.0",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.23",
    "langchain-google-genai>=2.0.10",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.1",
]
